requirements.txt
streamlit
openai
pandas
python-dotenv
azure-ai-documentintelligence